Reduced the double minus signs. Notice that the rols (y-axis) PID has negative gains.
GoT/position sensor convention: x-axis is positive in the FW direction of the quad, the y-axis is positive in the port/left side of the quad, the z-axis is positive upwards. The quad/Crazyflie commands are the following: pitch + is nose down (=fw on the x-position axis), pitch - is nose up (backwards on the x-position axis), roll + starboard/right (negative on the y-position axis), roll - port/left (positive on the y-position axis). For completness yaw + leads to clockwise spin, yaw - leads to counter-clockwise turn.